Remove complex parser and add some tests and fixed for charge parsing
Removed the chemical parser that can handle the middle-dot formulae, as I don't think it will ever be useful. The "basic" parser has now also been renamed to just parser due to this change.

Some sanity and correctness fixes to the way charges are handled, and added some tests to cover all of those cases.