Implement a Gauss Jordan RREF, clean up
Implement a Gauss Jordan RREF, no tests for this yet.

Implement the index trait, which will now return a row from the matrix. It can be subindexed to yield a specific entry for free since Vec<> supports indexing natively too. Switch the internal code to using that syntax and remove the get methods.